requirements.txt
jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
uWSGI==2.0.28
